requirements.txt
Flask
requests
nbformat
nbconvert
schedule
sendgrid
ipykernel
jupyter
ipython
python-dotenv
jupyter_client
google-generativeai
alpaca-py
alpaca